Delete TdrManifestImportMetadata class
